refactor(passkey): move background flows out of main controller
